refactor(architecture): refactor comments
