+++ router/comments/handlers/handleAddComment.js
@@ -1,0 +1,76 @@
+const { Upload, PhotoComment, User } = require('../../../models');
+const removeSpacesAndDashes = require('../../../utils/removeSpacesAndDashes');
+const normalizeS3Key = require('../../../utils/normalizeS3Key');
+
+const handleAddComment = async (req, res) => {
+  try {
+    const { uploadId, comment, taggedUserIds } = req.body;
+    const userId = req.auth.user.id;
+
+    const s3Key = req.file?.transforms?.[0]?.key ?? null;
+    let commentImageUpload = null;
+    let imageUrl = null;
+
+    if (s3Key) {
+      const cleanedName = removeSpacesAndDashes(
+        req.file.originalname.toLowerCase().trim()
+      );
+      const normalizedKey = normalizeS3Key(s3Key); // Removes env and sanitizes
+
+      commentImageUpload = await Upload.create({
+        url: s3Key,
+        name: cleanedName,
+        userId,
+      });
+
+      imageUrl = normalizedKey; // key without env prefix for PhotoComment
+    }
+
+    const photoComment = await PhotoComment.create({
+      userId,
+      uploadId,
+      comment,
+      imageUrl,
+    });
+
+    if (taggedUserIds && typeof taggedUserIds === 'string') {
+      const parsedTags = JSON.parse(taggedUserIds);
+      if (Array.isArray(parsedTags) && parsedTags.length > 0) {
+        await photoComment.setTaggedUsers(parsedTags);
+      }
+    }
+
+    const fullComment = await PhotoComment.findByPk(photoComment.id, {
+      include: [
+        { model: User, as: 'user', attributes: ['id', 'username'] },
+        { model: User, as: 'taggedUsers', attributes: ['id', 'username'] },
+      ],
+    });
+
+    const securePhotoUrl = imageUrl
+      ? `${process.env.API_BASE_URL}/uploads/files/${encodeURIComponent(
+          `${process.env.NODE_ENV}/${imageUrl}`
+        )}`
+      : null;
+
+    return res.status(201).send({
+      data: {
+        ...fullComment.toJSON(),
+        securePhotoUrl,
+      },
+    });
+  } catch (error) {
+    if (error.code === 'LIMIT_FILE_SIZE') {
+      return res.status(400).json({ message: 'Image too big' });
+    }
+
+    if (error.message?.includes('Invalid file type')) {
+      return res.status(400).json({ message: error.message });
+    }
+
+    console.error('❌ Error adding comment:', error);
+    return res.status(500).json({ message: 'Something went wrong' });
+  }
+};
+
+module.exports = handleAddComment;
