+++ router/comments/s3/uploadCommentImage.js
@@ -1,0 +1,41 @@
+const multer = require('multer');
+const multerS3 = require('multer-s3-transform');
+const sharp = require('sharp');
+const s3 = require('../../../utils/s3');
+const allowedMimeTypes = require("./../../../consts/allowedFileTypes");
+const removeSpacesAndDashes = require("../../../utils/removeSpacesAndDashes");
+
+const uploadCommentImage = multer({
+  storage: multerS3({
+    s3,
+    bucket: "duga-user-photo",
+    contentType: multerS3.AUTO_CONTENT_TYPE,
+    shouldTransform: true,
+    transforms: [
+      {
+        id: "commentImageResized",
+        key: function (req, file, cb) {
+          const timestamp = Date.now();
+          const cleanedFilename = removeSpacesAndDashes(file.originalname.toLowerCase().trim());
+          const path = `${process.env.NODE_ENV}/comment/${timestamp}/${cleanedFilename}`;
+          cb(null, path);
+        },
+        transform: function (req, file, cb) {
+          cb(null, sharp().resize(1024).jpeg({ quality: 80 }));
+        },
+      },
+    ],
+  }),
+  limits: { fileSize: 1 * 1024 * 1024 }, // 1MB
+  fileFilter: (req, file, cb) => {
+    if (allowedMimeTypes.includes(file.mimetype)) {
+      cb(null, true);
+    } else {
+      const error = new Error('Invalid file type. Only PNG, JPG, JPEG, and SVG are allowed.');
+      error.code = 'INVALID_FILE_TYPE';
+      cb(error);
+    }
+  },
+});
+
+module.exports = uploadCommentImage;
